# contract/openapi-1.2.0.json
{
  "components": {
    "schemas": {
...
      "Answer": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "answer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Id"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "audit_id": {
            "title": "Audit Id",
            "type": "string"
          },
          "contributing_sources": {
            "items": {
              "$ref": "#/components/schemas/ContributingSource"
            },
            "title": "Contributing Sources",
            "type": "array"
          },
          "drillthrough_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drillthrough Token"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "route": {
            "title": "Route",
            "type": "string"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count"
          },
          "rows": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows"
          },
          "sql_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sql Used"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "title": "Suggestions",
            "type": "array"
          }
        },
        "required": [
          "answer",
          "audit_id",
          "route",
          "provenance"
        ],
        "title": "Answer",
        "type": "object"
      },
...
      "AskRequest": {
        "properties": {
          "question": {
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "session_id": {
            "default": "demo",
            "title": "Session Id",
            "type": "string"
          },
          "space_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Space Id"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
...
      "Badge": {
        "enum": [
          "certified",
          "governed_metric",
          "query_skill",
          "session",
          "abstain",
          "blocked"
        ],
        "title": "Badge",
        "type": "string"
      },
...
      "ChainVerification": {
        "properties": {
          "broken_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broken At"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ChainVerification",
        "type": "object"
      },
...
      "ContributingSource": {
        "description": "One source card for the Sources panel (architecture \u00a74.7 / \u00a74.8).",
        "properties": {
          "container": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Container"
          },
          "contribution": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contribution"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "member": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member"
          },
          "ref_id": {
            "title": "Ref Id",
            "type": "string"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count"
          }
        },
        "required": [
          "ref_id"
        ],
        "title": "ContributingSource",
        "type": "object"
      },
...
      "DrillthroughRequest": {
        "properties": {
          "token": {
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "DrillthroughRequest",
        "type": "object"
      },
      "DrillthroughResponse": {
        "properties": {
          "answer_id": {
            "title": "Answer Id",
            "type": "string"
          },
          "approximate": {
            "default": false,
            "title": "Approximate",
            "type": "boolean"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "sql_used": {
            "title": "Sql Used",
            "type": "string"
          },
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count"
          }
        },
        "required": [
          "answer_id",
          "session_id",
          "sql_used",
          "row_count"
        ],
        "title": "DrillthroughResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LedgerAppendRequest": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "actor",
          "event_type"
        ],
        "title": "LedgerAppendRequest",
        "type": "object"
      },
      "LedgerEntry": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "entry_hash": {
            "title": "Entry Hash",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "prev_hash": {
            "title": "Prev Hash",
            "type": "string"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "seq",
          "actor",
          "event_type",
          "payload",
          "prev_hash",
          "entry_hash",
          "created_at"
        ],
        "title": "LedgerEntry",
        "type": "object"
      },
      "LedgerVerifyRequest": {
        "properties": {
          "start_seq": {
            "default": 0,
            "title": "Start Seq",
            "type": "integer"
          }
        },
        "title": "LedgerVerifyRequest",
        "type": "object"
      },
...
      "PoolSpec": {
        "properties": {
          "class_name": {
            "default": "default",
            "title": "Class Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "max_concurrency": {
            "default": 1,
            "title": "Max Concurrency",
            "type": "integer"
          }
        },
        "required": [
          "id"
        ],
        "title": "PoolSpec",
        "type": "object"
      },
...
      "Provenance": {
        "properties": {
          "assumptions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assumptions"
          },
          "badge": {
            "$ref": "#/components/schemas/Badge"
          },
          "layer": {
            "title": "Layer",
            "type": "string"
          },
          "metric_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric Id"
          },
          "query_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Source"
          }
        },
        "required": [
          "layer",
          "badge"
        ],
        "title": "Provenance",
        "type": "object"
      },
...
      "QueryResult": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "output": {
            "title": "Output"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "ok",
          "status"
        ],
        "title": "QueryResult",
        "type": "object"
      },
...
      "SubmitRequest": {
        "properties": {
          "body": {
            "additionalProperties": true,
            "title": "Body",
            "type": "object"
          },
          "manifest": {
            "$ref": "#/components/schemas/Manifest"
          },
          "plan": {
            "additionalProperties": true,
            "title": "Plan",
            "type": "object"
          },
          "pool": {
            "$ref": "#/components/schemas/PoolSpec"
          }
        },
        "required": [
          "pool",
          "plan",
          "body",
          "manifest"
        ],
        "title": "SubmitRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Allowlisted DMS\u2194engine contract surface (operationIds: ask, drillthrough, ledger.append, ledger.verify, submit, tool.registry). cortex_contract 1.2.0. Engine semver is independent (see CortexOS.__version__ / docs/RELEASING.md).",
    "title": "Cortex Engine Contract API",
    "version": "1.2.0",
    "x-cortex-contract-routes": [
      "ask",
      "drillthrough",
      "ledger.append",
      "ledger.verify",
      "submit",
      "tool.registry"
    ],
    "x-cortex-contract-version": "1.2.0",
    "x-cortex-ports": [
      8000,
      8010
    ]
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/contract/ask": {
      "post": {
        "description": "Answer a governed question (DMS ask plane). Requires a prior submit bind.",
        "operationId": "ask",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Contract Ask",
        "tags": [
          "contract"
        ]
      }
    },
    "/v1/contract/drillthrough": {
      "post": {
        "description": "Re-derive contributing rows under the same session manifest (T7).",
        "operationId": "drillthrough",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DrillthroughRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrillthroughResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Contract Drillthrough",
        "tags": [
          "contract"
        ]
      }
    },
    "/v1/contract/ledger/append": {
      "post": {
        "operationId": "ledger.append",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LedgerAppendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Contract Ledger Append",
        "tags": [
          "contract"
        ]
      }
    },
    "/v1/contract/ledger/verify": {
      "post": {
        "operationId": "ledger.verify",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LedgerVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainVerification"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Contract Ledger Verify",
        "tags": [
          "contract"
        ]
      }
    },
    "/v1/contract/submit": {
      "post": {
        "description": "Submit a plan under a signed session manifest (C4).",
        "operationId": "submit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Contract Submit",
        "tags": [
          "contract"
        ]
      }
    },
...
  }
}